# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sokunthy Vathana",
        "description": "API documentation for managing products, branches, and more.",
        "contact": {
            "email": "sokunthyvathana@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/branches": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get paginated list of branches",
                "description": "Returns a paginated list of branches with optional query parameters for page size and limit.",
                "operationId": "8512c7cd0f421159fb7acb2f79570002",
                "parameters": [
                    {
                        "name": "_pageLimit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful pagination response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_items": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "branch_name": {
                                                        "type": "string",
                                                        "example": "Phnom Penh Branch"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Phnom Penh"
                                                    },
                                                    "contact_number": {
                                                        "type": "string",
                                                        "example": "012345678"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branch/{id}": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get branch by ID",
                "description": "Retrieve a single branch by its ID.",
                "operationId": "6949da17051fa360e5a81175fcef66f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the branch to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "branch_name": {
                                                    "type": "string",
                                                    "example": "Phnom Penh Branch"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Phnom Penh"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "012345678"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Branch with ID 99 not found"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branch/create": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Create a new branch",
                "description": "Creates a new branch with the given details.",
                "operationId": "ef83e9da50d4d2ff284385b83fd368af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_name",
                                    "location",
                                    "contact_number"
                                ],
                                "properties": {
                                    "branch_name": {
                                        "type": "string",
                                        "example": "Phnom Penh Branch"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Phnom Penh"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "012345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "new_branch": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "branch_name": {
                                                    "type": "string",
                                                    "example": "Phnom Penh Branch"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Phnom Penh"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "012345678"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "branch_name": {
                                                    "type": "string",
                                                    "example": "Branch name is not allowed to be null."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}